+++ modules/rank-eval/src/main/java/org/elasticsearch/index/rankeval/EvalQueryQuality.java
@@ -11,13 +11,10 @@ package org.elasticsearch.index.rankeval;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.io.stream.Writeable;
-import org.elasticsearch.common.xcontent.XContentParserUtils;
 import org.elasticsearch.index.rankeval.RatedDocument.DocumentKey;
-import org.elasticsearch.xcontent.ObjectParser;
 import org.elasticsearch.xcontent.ParseField;
 import org.elasticsearch.xcontent.ToXContentFragment;
 import org.elasticsearch.xcontent.XContentBuilder;
-import org.elasticsearch.xcontent.XContentParser;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -35,24 +32,23 @@ public class EvalQueryQuality implements ToXContentFragment, Writeable {
     private final List<RatedSearchHit> ratedHits;
 
     public EvalQueryQuality(String id, double metricScore) {
-        this.queryId = id;
-        this.metricScore = metricScore;
-        this.ratedHits = new ArrayList<>();
+        this(id, metricScore, new ArrayList<>(), null);
     }
 
     public EvalQueryQuality(StreamInput in) throws IOException {
-        this.queryId = in.readString();
-        this.metricScore = in.readDouble();
-        this.ratedHits = in.readCollectionAsList(RatedSearchHit::new);
-        this.optionalMetricDetails = in.readOptionalNamedWriteable(MetricDetail.class);
+        this(
+            in.readString(),
+            in.readDouble(),
+            in.readCollectionAsList(RatedSearchHit::new),
+            in.readOptionalNamedWriteable(MetricDetail.class)
+        );
     }
 
-    // only used for parsing internally
-    private EvalQueryQuality(String queryId, ParsedEvalQueryQuality builder) {
+    EvalQueryQuality(String queryId, double evaluationResult, List<RatedSearchHit> ratedHits, MetricDetail optionalMetricDetails) {
         this.queryId = queryId;
-        this.metricScore = builder.evaluationResult;
-        this.optionalMetricDetails = builder.optionalMetricDetails;
-        this.ratedHits = builder.ratedHits;
+        this.metricScore = evaluationResult;
+        this.optionalMetricDetails = optionalMetricDetails;
+        this.ratedHits = ratedHits;
     }
 
     @Override
@@ -113,37 +109,8 @@ public class EvalQueryQuality implements ToXContentFragment, Writeable {
 
     static final ParseField METRIC_SCORE_FIELD = new ParseField("metric_score");
     private static final ParseField UNRATED_DOCS_FIELD = new ParseField("unrated_docs");
-    private static final ParseField HITS_FIELD = new ParseField("hits");
-    private static final ParseField METRIC_DETAILS_FIELD = new ParseField("metric_details");
-    private static final ObjectParser<ParsedEvalQueryQuality, Void> PARSER = new ObjectParser<>(
-        "eval_query_quality",
-        true,
-        ParsedEvalQueryQuality::new
-    );
-
-    private static class ParsedEvalQueryQuality {
-        double evaluationResult;
-        MetricDetail optionalMetricDetails;
-        List<RatedSearchHit> ratedHits = new ArrayList<>();
-    }
-
-    static {
-        PARSER.declareDouble((obj, value) -> obj.evaluationResult = value, METRIC_SCORE_FIELD);
-        PARSER.declareObject((obj, value) -> obj.optionalMetricDetails = value, (p, c) -> parseMetricDetail(p), METRIC_DETAILS_FIELD);
-        PARSER.declareObjectArray((obj, list) -> obj.ratedHits = list, (p, c) -> RatedSearchHit.parse(p), HITS_FIELD);
-    }
-
-    private static MetricDetail parseMetricDetail(XContentParser parser) throws IOException {
-        XContentParserUtils.ensureExpectedToken(XContentParser.Token.START_OBJECT, parser.currentToken(), parser);
-        XContentParserUtils.ensureExpectedToken(XContentParser.Token.FIELD_NAME, parser.nextToken(), parser);
-        MetricDetail metricDetail = parser.namedObject(MetricDetail.class, parser.currentName(), null);
-        XContentParserUtils.ensureExpectedToken(XContentParser.Token.END_OBJECT, parser.nextToken(), parser);
-        return metricDetail;
-    }
-
-    public static EvalQueryQuality fromXContent(XContentParser parser, String queryId) throws IOException {
-        return new EvalQueryQuality(queryId, PARSER.apply(parser, null));
-    }
+    static final ParseField HITS_FIELD = new ParseField("hits");
+    static final ParseField METRIC_DETAILS_FIELD = new ParseField("metric_details");
 
     @Override
     public final boolean equals(Object obj) {
