+++ server/src/main/java/org/elasticsearch/action/search/MultiSearchResponse.java
@@ -24,35 +24,15 @@ import org.elasticsearch.core.Nullable;
 import org.elasticsearch.core.RefCounted;
 import org.elasticsearch.core.TimeValue;
 import org.elasticsearch.transport.LeakTracker;
-import org.elasticsearch.xcontent.ConstructingObjectParser;
-import org.elasticsearch.xcontent.ParseField;
 import org.elasticsearch.xcontent.ToXContent;
-import org.elasticsearch.xcontent.XContentParser;
-import org.elasticsearch.xcontent.XContentParser.Token;
 
 import java.io.IOException;
 import java.util.Iterator;
-import java.util.List;
-
-import static org.elasticsearch.xcontent.ConstructingObjectParser.constructorArg;
 
 /**
  * A multi search response.
  */
 public class MultiSearchResponse extends ActionResponse implements Iterable<MultiSearchResponse.Item>, ChunkedToXContentObject {
-
-    private static final ParseField RESPONSES = new ParseField(Fields.RESPONSES);
-    private static final ParseField TOOK_IN_MILLIS = new ParseField("took");
-    @SuppressWarnings("unchecked")
-    private static final ConstructingObjectParser<MultiSearchResponse, Void> PARSER = new ConstructingObjectParser<>(
-        "multi_search",
-        true,
-        a -> new MultiSearchResponse(((List<Item>) a[0]).toArray(new Item[0]), (long) a[1])
-    );
-    static {
-        PARSER.declareObjectArray(constructorArg(), (p, c) -> itemFromXContent(p), RESPONSES);
-        PARSER.declareLong(constructorArg(), TOOK_IN_MILLIS);
-    }
 
     /**
      * A search response item, holding the actual search response, or an error message if it failed.
@@ -231,47 +211,8 @@ public class MultiSearchResponse extends ActionResponse implements Iterable<Mult
         );
     }
 
-    public static MultiSearchResponse fromXContext(XContentParser parser) {
-        return PARSER.apply(parser, null);
-    }
-
-    private static MultiSearchResponse.Item itemFromXContent(XContentParser parser) throws IOException {
-        // This parsing logic is a bit tricky here, because the multi search response itself is tricky:
-        // 1) The json objects inside the responses array are either a search response or a serialized exception
-        // 2) Each response json object gets a status field injected that ElasticsearchException.failureFromXContent(...) does not parse,
-        // but SearchResponse.innerFromXContent(...) parses and then ignores. The status field is not needed to parse
-        // the response item. However in both cases this method does need to parse the 'status' field otherwise the parsing of
-        // the response item in the next json array element will fail due to parsing errors.
-
-        Item item = null;
-        String fieldName = null;
-
-        Token token = parser.nextToken();
-        assert token == Token.FIELD_NAME;
-        outer: for (; token != Token.END_OBJECT; token = parser.nextToken()) {
-            switch (token) {
-                case FIELD_NAME:
-                    fieldName = parser.currentName();
-                    if ("error".equals(fieldName)) {
-                        item = new Item(null, ElasticsearchException.failureFromXContent(parser));
-                    } else if ("status".equals(fieldName) == false) {
-                        item = new Item(SearchResponse.innerFromXContent(parser), null);
-                        break outer;
-                    }
-                    break;
-                case VALUE_NUMBER:
-                    if ("status".equals(fieldName)) {
-                        // Ignore the status value
-                    }
-                    break;
-            }
-        }
-        assert parser.currentToken() == Token.END_OBJECT;
-        return item;
-    }
-
-    static final class Fields {
-        static final String RESPONSES = "responses";
+    public static final class Fields {
+        public static final String RESPONSES = "responses";
         static final String STATUS = "status";
     }
