+++ x-pack/plugin/security/qa/multi-cluster/src/javaRestTest/java/org/elasticsearch/xpack/remotecluster/AbstractRemoteClusterSecurityWithMultipleRemotesRestIT.java
@@ -16,6 +16,7 @@ import org.elasticsearch.client.RestClient;
 import org.elasticsearch.core.IOUtils;
 import org.elasticsearch.core.Strings;
 import org.elasticsearch.search.SearchHit;
+import org.elasticsearch.search.SearchResponseUtils;
 import org.elasticsearch.test.cluster.ElasticsearchCluster;
 import org.elasticsearch.xcontent.ObjectPath;
 import org.junit.AfterClass;
@@ -216,7 +217,7 @@ public abstract class AbstractRemoteClusterSecurityWithMultipleRemotesRestIT ext
         assertOK(response);
         final SearchResponse searchResponse;
         try (var parser = responseAsParser(response)) {
-            searchResponse = SearchResponse.fromXContent(parser);
+            searchResponse = SearchResponseUtils.parseSearchResponse(parser);
         }
         try {
             final List<String> actualIndices = Arrays.stream(searchResponse.getHits().getHits())
