+++ x-pack/plugin/security/qa/multi-cluster/src/javaRestTest/java/org/elasticsearch/xpack/remotecluster/RemoteClusterSecurityCcrIT.java
@@ -14,6 +14,7 @@ import org.elasticsearch.client.Response;
 import org.elasticsearch.client.ResponseException;
 import org.elasticsearch.core.Strings;
 import org.elasticsearch.search.SearchHit;
+import org.elasticsearch.search.SearchResponseUtils;
 import org.elasticsearch.test.cluster.ElasticsearchCluster;
 import org.elasticsearch.test.cluster.util.resource.Resource;
 import org.elasticsearch.test.rest.ObjectPath;
@@ -278,7 +279,7 @@ public class RemoteClusterSecurityCcrIT extends AbstractRemoteClusterSecurityTes
             assertOK(response);
             final SearchResponse searchResponse;
             try (var parser = responseAsParser(response)) {
-                searchResponse = SearchResponse.fromXContent(parser);
+                searchResponse = SearchResponseUtils.parseSearchResponse(parser);
             }
             try {
                 assertThat(searchResponse.getHits().getTotalHits().value, equalTo(numberOfDocs));
