+++ x-pack/plugin/security/qa/multi-cluster/src/javaRestTest/java/org/elasticsearch/xpack/remotecluster/RemoteClusterSecuritySpecialUserIT.java
@@ -14,6 +14,7 @@ import org.elasticsearch.client.Response;
 import org.elasticsearch.client.ResponseException;
 import org.elasticsearch.core.Strings;
 import org.elasticsearch.search.SearchHit;
+import org.elasticsearch.search.SearchResponseUtils;
 import org.elasticsearch.test.cluster.ElasticsearchCluster;
 import org.elasticsearch.test.cluster.util.resource.Resource;
 import org.elasticsearch.test.junit.RunnableTestRuleAdapter;
@@ -137,7 +138,7 @@ public class RemoteClusterSecuritySpecialUserIT extends AbstractRemoteClusterSec
                 new Request("GET", "/my_remote_cluster:" + randomFrom("*", "shared-*", "shared-logs") + "/_search")
             );
             assertOK(response1);
-            final SearchResponse searchResponse1 = SearchResponse.fromXContent(responseAsParser(response1));
+            final SearchResponse searchResponse1 = SearchResponseUtils.parseSearchResponse(responseAsParser(response1));
             try {
                 assertThat(
                     Arrays.stream(searchResponse1.getHits().getHits()).map(SearchHit::getIndex).collect(Collectors.toList()),
@@ -181,7 +182,7 @@ public class RemoteClusterSecuritySpecialUserIT extends AbstractRemoteClusterSec
                 )
             );
             assertOK(response3);
-            final SearchResponse searchResponse3 = SearchResponse.fromXContent(responseAsParser(response3));
+            final SearchResponse searchResponse3 = SearchResponseUtils.parseSearchResponse(responseAsParser(response3));
             try {
                 assertThat(
                     Arrays.stream(searchResponse3.getHits().getHits()).map(SearchHit::getIndex).collect(Collectors.toList()),
@@ -201,7 +202,7 @@ public class RemoteClusterSecuritySpecialUserIT extends AbstractRemoteClusterSec
             kibanaServiceSearchRequest.setOptions(RequestOptions.DEFAULT.toBuilder().addHeader("Authorization", "Bearer " + serviceToken));
             final Response kibanaServiceSearchResponse = client().performRequest(kibanaServiceSearchRequest);
             assertOK(kibanaServiceSearchResponse);
-            final SearchResponse searchResponse4 = SearchResponse.fromXContent(responseAsParser(kibanaServiceSearchResponse));
+            final SearchResponse searchResponse4 = SearchResponseUtils.parseSearchResponse(responseAsParser(kibanaServiceSearchResponse));
             try {
                 assertThat(
                     Arrays.stream(searchResponse4.getHits().getHits()).map(SearchHit::getIndex).collect(Collectors.toList()),
@@ -223,7 +224,7 @@ public class RemoteClusterSecuritySpecialUserIT extends AbstractRemoteClusterSec
             );
             final Response elasticUserSearchResponse = client().performRequest(elasticUserSearchRequest);
             assertOK(elasticUserSearchResponse);
-            final SearchResponse searchResponse5 = SearchResponse.fromXContent(responseAsParser(elasticUserSearchResponse));
+            final SearchResponse searchResponse5 = SearchResponseUtils.parseSearchResponse(responseAsParser(elasticUserSearchResponse));
             try {
                 assertThat(
                     Arrays.stream(searchResponse5.getHits().getHits()).map(SearchHit::getIndex).collect(Collectors.toList()),
