+++ x-pack/plugin/security/qa/multi-cluster/src/javaRestTest/java/org/elasticsearch/xpack/remotecluster/RemoteClusterSecurityTopologyRestIT.java
@@ -13,6 +13,7 @@ import org.elasticsearch.client.RequestOptions;
 import org.elasticsearch.client.Response;
 import org.elasticsearch.core.Strings;
 import org.elasticsearch.search.SearchHit;
+import org.elasticsearch.search.SearchResponseUtils;
 import org.elasticsearch.test.cluster.ElasticsearchCluster;
 import org.elasticsearch.test.cluster.util.resource.Resource;
 import org.elasticsearch.test.junit.RunnableTestRuleAdapter;
@@ -127,7 +128,7 @@ public class RemoteClusterSecurityTopologyRestIT extends AbstractRemoteClusterSe
         {
             final var documentFieldValues = new HashSet<>();
             final var searchRequest = new Request("GET", "/my_remote_cluster:*/_search?scroll=1h&size=1");
-            final SearchResponse searchResponse = SearchResponse.fromXContent(
+            final SearchResponse searchResponse = SearchResponseUtils.parseSearchResponse(
                 responseAsParser(performRequestWithRemoteMetricUser(searchRequest))
             );
             final Request scrollRequest = new Request("GET", "/_search/scroll");
@@ -148,7 +149,7 @@ public class RemoteClusterSecurityTopologyRestIT extends AbstractRemoteClusterSe
 
             // Fetch all documents
             for (int i = 0; i < 5; i++) {
-                final SearchResponse scrollResponse = SearchResponse.fromXContent(
+                final SearchResponse scrollResponse = SearchResponseUtils.parseSearchResponse(
                     responseAsParser(performRequestWithRemoteMetricUser(scrollRequest))
                 );
                 try {
